[Feat] #23 - 팝업 테스트 메뉴 구성
